requirements/dev.txt
-r base.txt

black>=24.0
flake8>=7.0
pytest>=8.0
pytest-django>=4.8
django-debug-toolbar>=4.4

